Add open source license
Add a new license file using the default apache 2.0 template. I know Deepspeech uses MPL however after looking into it, I think we are fine to keep using Apache as long as we keep the MPL license on the Deepspeech codebase (and since we aren't modifying it, we should be fine).